\name{SMCMP-package}
\alias{SMCMP-package}
\alias{SMCMP}
\docType{package}
\title{
\packageTitle{SMCMP}
}
\description{
\packageDescription{SMCMP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SMCMP}
\packageIndices{SMCMP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SMCMP}

Maintainer: \packageMaintainer{SMCMP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
